==> Development/1_FirstCleaning.ipynb <==
# Linkedin Posts Analysis

## Question : what makes a post good ?

Nowadays, a lot of businesses are based on the relationship with their prospects and clients.   
Moreover, a lot of "Linkedin experts" propose their services for companies to increase their visibility on this social media.  

Understanding what makes a post good on LinkedIn is essential to increase the reach, interactions with prospects, to create trust and therefore increase business revenus. 

In this notebook, we will study a dataset provided on Kaggle to understand which criteria make a post really "good" on this social media. 

We call a "good post" content which creates interactions with the community. 

### Datasets comments on Kaggle

This dataset contains LinkedIn Influencers' post details and other details(post dependent as well as independent) per post. This dataset can be used to analyze LinkedIn reach based on post content and related account details.

This dataset is great for Exploratory Data Analysis and NLP tasks.

The data was scraped using BeautifulSoup and Selenium.Last updated on 15th Feb,2021

### Principal steps of this analysis
* Understand the dataset
* Cleaning
* EDA

# Libraries

import pandas as pd
import matplotlib.pyplot as plt
import requests
import io

# Downloading the dataset

# # Downloading the csv.file from my Github account

# # We copy the "raw" link
# url = "https://github.com/JeremyArancio/Linkedin_Post_Analysis/blob/main/influencers_data.csv?raw=true"
# download = requests.get(url).content

# # Reading the downloaded content and turning it into a pandas dataframe
# df = pd.read_csv(io.StringIO(download.decode('utf-8')))

df = pd.read_csv('influencers_data.csv')

# First look

df.info()

#Nan in %
df.isna().sum() / df.shape[0] * 100

df.head(1)

## Observations for the Comments section

#Let's check Time_spent
df.time_spent.value_counts().head(3)

#Let's check Media_type
df.media_type.value_counts()

#Content_links
df.content_links[0]

#Media_url
df.media_url.head()[0]

#Hashtags_followers
df.hashtag_followers.value_counts()

#hashtags
df.hashtags[0]

df.votes.value_counts()

#Connections
df.connections.value_counts()

#Location
df.location.value_counts().head()

#Location nan
df.location.isna().sum()

#Media types
df.media_type.value_counts()

#Media types
df.media_type.isna().sum()

## Comments

We have different columns :
* Unnamed : to delete
* Name : Linkedin profile
* Headline : below the name in Linkedin
* Location : some Nan ; List : example ['Gloucester,', 'Massachusetts,', 'United', 'States']
* Followers : number of followers (connextions + followers in Linkedin) ; This dataset is composed of influencers with a lot of followers
* Connections : careful of 500+ (is it useful ?) => only 2 differents values => delete it !
* About : about section in the profile
* time_spent : time between the post date & when this dataset was created (categorical values : 1 month ago, 5 years ago, ...)
* Content : core of the content
* Content_links : every links conatined in the post : profile, #, media (website, picture,...)
* Media_type : media added to the post : article, image, photo,... ; Some nan
* Media_url : link of the media
* Num_hashtags : number
* Hashtag_followers : nulls
* Hashtags : [['#',"link"]]
* Reactions : number of likes, supports, loves,... counted as identical
* Comments : number of comments
* Views : Nan
* Votes : 82 non-null, composed of decimal numbers too. Weird. Because it is unsignificant, we could delete it

# Cleaning

## Drop useless columns

df_dropped = df.drop(['Unnamed: 0','connections','hashtag_followers','views','votes'],axis=1)

## Replace Nan

New save of the dataset modified

df_fillNa = df_dropped.copy()

df_fillNa.isna().sum()

### Location

Some locations are missing, then we are going to replace them by Unknown

#We add a new column 'Locations'
df_fillNa["Locations"] = df_dropped['location'].fillna('Unknown')

#Then we drop the previous "location" column
df_fillNa = df_fillNa.drop(['location'],axis=1)

### Followers

To avoid any bias, let's take the mean to replace Nan.  
There are only 42 missing values.

Followers_mean = df_fillNa.followers.mean()
Followers_mean

df_fillNa["Followers"] = df_fillNa['followers'].fillna(Followers_mean)

#Drop
df_fillNa = df_fillNa.drop(['followers'],axis=1)

### Time spent

There is only one missing value. 

df_fillNa.loc[df_fillNa['time_spent'].isna()]

We see it's about coronavirus.  
We can think the post was published the day the extraction was done.  
For this reason, let's replace Nan by "Current date"

df_fillNa['Time_spent'] = df_fillNa['time_spent'].fillna('Current date')

#Drop the previous column
df_fillNa = df_fillNa.drop(['time_spent'],axis=1)

### Content

The question we decided to treat is about post contents.  
Therefore we have no reason to keep rows without content.  
Let's drop them

df_fillNa = df_fillNa.loc[df_fillNa['content'].notnull()]

### Media type

We suppose that Nan means there is no media in the post.
Let's replace it by "None"

df_fillNa["Media_type"] = df_fillNa['media_type'].fillna("none")

df_fillNa = df_fillNa.drop(['media_type'],axis=1)

### Check if everything is ok

df_fillNa.isna().sum()

df_fillNa.shape

print('We dropped {} rows from the original dataset'.format(df.shape[0]-df_fillNa.shape[0]))

The data is now ready to be explored.

# Export the cleaned Dataset

df_cleaned = df_fillNa.copy()

df_cleaned.info()

import pickle
df_cleaned.to_pickle('cleaned_data.pkl')

Let's go to the next notebook
