==> Development/2_NLPCleaning.ipynb <==
# NLP Cleaning

In this notebook, we will adapt and modify the cleaned dataset to apply Natural Language Processing on it.  
Some columns are composed of texts we can transform to obtain an exploitable dataset. 

## Libraries

import pickle
import pandas as pd
import warnings
warnings.filterwarnings('ignore') #Remove warnings

## Import the data

data = pd.read_pickle('cleaned_data.pkl')

## First look

### Overall

data.head(3)

### Content

data.content[1000]

**Comments**  
We see "...see more" in the content, which means we don't have all the content !  
Let's see if we can manage to get the entire content with the links provided in the dataset. 

### About

data.about[1000]

### Headline

print(data.headline[1000])
print(data.headline[100])
print(data.headline[10000])

**Comments**  
With headlines, maybe we can highlight the postion of the person.  
Also we can analyse which words or expressions stand out ("help businesses" ; "Start-up"  ...)

### Content links

print(data.content_links[100])
print("--------------")
print(data.content_links[10000])

**Comments**  
There are 3 types of links :
1. Other account/people mentions
2. Hastags
3. Link to websites

The data is organized as the following :
[[ "Link" , "mention/hashtag/website_link" ]]

We can differenciate these 3 types depending easily. For instance :  
1. Capital at the beginning of the string
2. "#"
3. "http"

### Media URL

print(data.media_url[10000])
print('-----------')
print(data.content_links[10000])

#Let's see the composition of media_url
data.media_url.value_counts().head()

#We observe some [] in it, which means that some contents (a lot actually) don't have media


**Comments**  
When a media is added to the content, an option allowed by Linkedin, the link is added to this column.   

We observe the difference between "content_links" & "media_url" here :  
* Content_links gathers all links inside the writen content. 
* Media_url gathers all the links of the added media, if there is one


### Hashtags

print(data.hashtags[10000])
print("------------")
print(data.content_links[10000])

**Comments**  
We observe that links in the content_links are also in hashtags column.  
Thus, this information appears two times in the dataset. 

## Conclusions on the first look

To wrap it up :

* Contents are not complete. A new scraping should be recquired to gather more data. **However**, we can use the first sentences to draw some insights. Indeed, the tagline is usually more important than the core of the content. 
* Headlines contain a lot of information, like position or business
* Content_links contains links of mentions/hashtags/websites used in the post. Hashtags are contained also in its own column. 
* Media_URL is present if a media is added to the post. 
* Time spent column could be changed into days to make easier the analysis.
* "About" section is complete and is a nice way to know in detail the person.

data.head(2)

## Analysis ideas

1. Study the content as it is. Even if it is not complete, we can considere in our analysis the first sentences present in this dataset. NLP technique can be easily applied here.
2. NLP techniques applied to headlines. 
3. Count mentions/hashtags in posts
4. Influence of media presence (media or not : 1-0)
5. NLP to About section
6. Relation between number of followers & reactions/comments
7. Influence of the media type

Reactions & Comments will be the main insight to define if a post was good or not. 

## Datasets creation

For each analysis, we are going to create adapted datasets and do some cleaning of the original one.

**Cleaning The Data**    

When dealing with numerical data, data cleaning often involves removing null values and duplicate data, dealing with outliers, etc. With text data, there are some common data cleaning techniques, which are also known as text pre-processing techniques.

With text data, this cleaning process can go on forever. There's always an exception to every cleaning step. So, we're going to follow the MVP (minimum viable product) approach - start simple and iterate. Here are a bunch of things you can do to clean your data. We're going to execute just the common cleaning steps here and the rest can be done at a later point to improve our results.

**Common data cleaning steps on all text:**

* Make text all lower case  
* Remove punctuation
* Remove numerical values
* Remove common non-sensical text (/n)
* Tokenize text
* Remove stop words

**More data cleaning steps after tokenization:**

* Stemming / lemmatization
* Parts of speech tagging
* Create bi-grams or tri-grams
* Deal with typos
* And more...

### Content dataset

For the content, we will apply NLP techniques.  
To do so, we need a **Corpus** & a **Document-term matrix**

A **Corpus** is a collection of texts.

A **Document-term matrix** is a dataframe added with a column for each word, and inside it, the number of times this word appears in the text

#### Corpus

contentCorpus = pd.DataFrame()
contentCorpus[['Name','Content','#Reactions','#Comments','Location','Followers','Time_spent','Media_type']] = data[['name','content','reactions', 'comments', 'Locations','Followers', 'Time_spent', 'Media_type']]
contentCorpus.head()

**Cleaning**  
Let's remove :
* Punctuation
* Lowercase letters
* Remove numbers
* Remove # (hashtags) => we will just remove # because some authors decided to incorporate #word inside their content
* Remove "/n/n/n see more..."

import re
import string

def clean_text(text):
    '''Make text lowercase, remove text in square brackets, remove punctuation, 
        remove words containing numbers, remove # and remove "see more" at the end of the LinkedIn post'''
    text = text.lower()
    text = re.sub('\[.*?\] ', '', text)
    text = re.sub('[%s]' % re.escape(string.punctuation), '', text)
    text = re.sub('\w*\d\w*', '', text)
    text = re.sub('#','',text)
    text = re.sub("\n \n \n …see more",'',text)
    return text

cleaning_text = lambda x: clean_text(x)

#Clean the content and create a new column              
contentCorpus['CleanedContent'] = contentCorpus['Content'].apply(cleaning_text)   

#Drop the previous column "Content"
contentCorpus = contentCorpus.drop(['Content'],axis=1)
contentCorpus.head(2)

#Let's see some examples
print(contentCorpus.CleanedContent[100])
print('---------')
print(contentCorpus.CleanedContent[1000])
print('---------')
print(contentCorpus.CleanedContent[10000])

We note that some names are present in the text.  
Nevertheless, the Corpus can be exploited right now for a *first attempt*

#Let's rename CleanedContent into Content for better understanding
contentCorpus = contentCorpus.rename({"CleanedContent":"Content"},axis=1)
contentCorpus

#Let's pickle the dataset for later use
contentCorpus.to_pickle("contentCorpus.pkl")

#### Document-term matrix

For many of the techniques we'll be using in future notebooks, the text must be tokenized, meaning broken down into smaller pieces. The most common tokenization technique is to break down text into words. We can do this using scikit-learn's CountVectorizer, where every row will represent a different document and every column will represent a different word.

In addition, with CountVectorizer, we can remove stop words. Stop words are common words that add no additional meaning to text such as 'a', 'the', etc.

*Memory issue*

It appears that creating a Document-term matrix of all the content is impossible due to the weight of the resulting dataset. To avoid this issue, we are going to take only few authors to create this dataset.  
Moreover, in order to analyze what makes a linkedin post good, it would be better to compare posts of the same author only.  
Let's take Simon Sinek, a famous personality in Personal Development, writer and public speaker.


#Corpus of Simon Sinek's content
sinekCorpus = contentCorpus.loc[contentCorpus["Name"] == "Simon Sinek"]
#Reset index to allow concat with dataframe
sinekCorpus = sinekCorpus.reset_index(drop=True) #drop to avoid current index into the dataframe 

print(sinekCorpus.shape)
sinekCorpus.head(2)

#Drop useless columns for this analysis too.
sinekCorpus.drop(['Name','Location','Time_spent','Media_type','Followers'],axis=1,inplace=True)

# We are going to create a document-term matrix using CountVectorizer, and exclude common English stop words
from sklearn.feature_extraction.text import CountVectorizer

cv = CountVectorizer(stop_words='english')
dataCv = cv.fit_transform(sinekCorpus.Content)
dataDtm = pd.DataFrame(dataCv.toarray(), columns=cv.get_feature_names_out())

#Pickle cv for later use
pickle.dump(cv, open("cv_stop.pkl", "wb"))

print(dataDtm.shape)
dataDtm.head(3)

#Let's concat this Document-term matrix with data to perform EDA
sinekDtm = pd.concat([sinekCorpus,dataDtm],axis=1)

print(sinekDtm.shape)
sinekDtm.head(2)

#Let's picke the Document-term matrix
sinekDtm.to_pickle("sinekDtm.pkl")

**Corpus for each author**

We will also study each creator content, and compare it with the number of followers for instance.

We can create a corpus for every authors.  
To do so, let's gather all content posts into one text for each author.  
Then we will apply CountVectorizer for each author.  
Easy !

contentAuthorCorpus = contentCorpus[['Name','Followers','Content']]

#Let's gather the content for each author
#We replace the 'Content' column
contentAuthorCorpus['Content'] = contentAuthorCorpus.groupby(['Name'])['Content'].transform(lambda x : ' '.join(x))

#Groupby group like a pivot_table
#Transform is Apply to the Serie itself
#Join gathers string together with the string ' ' defined (here nothing)

#Remove duplicates due to .transform
contentAuthorCorpus = contentAuthorCorpus.drop_duplicates() 
contentAuthorCorpus.reset_index(inplace=True,drop=True)
contentAuthorCorpus

#Export the dataset
contentAuthorCorpus.to_pickle('contentAuthorCorpus.pkl')

**Document-Term Matrix for headlines**
