==> Development/5_TopicModeling.ipynb <==
# Topic Modeling

A popular text analysis technique is called topic modeling. The ultimate goal of topic modeling is to find various topics that are present in the corpus. Each document in the corpus will be made up of at least one topic, if not multiple topics.

In this notebook, we will be covering the steps on how to do **Latent Dirichlet Allocation (LDA)**, which is one of many topic modeling techniques. It was specifically designed for text data.

To use a topic modeling technique, you need to provide (1) a document-term matrix and (2) the number of topics you would like the algorithm to pick up.

Once the topic modeling technique is applied, your job as a human is to interpret the results and see if the mix of words in each topic make sense. If they don't make sense, you can try changing up the number of topics, the terms in the document-term matrix, model parameters, or even try a different model.

## Libraries

import pandas as pd
import pickle
from gensim import matutils, models
import scipy.sparse

## LDA : 1st attempt

#Import the Document-term matrix of Simon Sinek
tdm = pd.read_pickle("sinekDtm.pkl")

#Let's put aside the content of simon sinek
sinekCorpus = tdm.Content

#We don't need the content, only the dtm
tdm.drop("Content",axis=1,inplace=True)

tdm.head()

#We remove Reactions & Comments from the Dtm to be used by the algorithm
tdmNoReactions = tdm.drop(['#Reactions','#Comments'],axis=1)

#We transpose the data for gensim
tdmNoReactions = tdmNoReactions.transpose()
tdmNoReactions

# We're going to put the term-document matrix into a new gensim format, from df --> sparse matrix --> gensim corpus
sparse_counts = scipy.sparse.csr_matrix(tdm)
corpus = matutils.Sparse2Corpus(sparse_counts) #It's actually a dtm

# Gensim also requires dictionary of the all terms and their respective location in the term-document matrix
cv = pickle.load(open("cv_stop.pkl", "rb"))
id2word = dict((v, k) for k, v in cv.vocabulary_.items())

Now that we have the corpus (term-document matrix) and id2word (dictionary of location: term), we need to specify two other parameters - the number of topics and the number of passes. Let's start the number of topics at 2, see if the results make sense, and increase the number from there.

# Now that we have the corpus (term-document matrix) and id2word (dictionary of location: term),
# we need to specify two other parameters as well - the number of topics and the number of passes
lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=4, passes=10)
lda.print_topics()

We see some words are unconvenient for this analyse (applecosimonsinek, actually, asked).  
Therefore we can clean the data in a better way.  

For instance we can :
* keep nouns & adjectives only
* remove the words most used

## 2nd attempt

#Let's use content corpus of Simon Sinek's posts we created above
sinekCorpus

One popular trick is to look only at terms that are from one part of speech (only nouns, only adjectives, etc.). Check out the UPenn tag set: https://www.ling.upenn.edu/courses/Fall_2003/ling001/penn_treebank_pos.html.

#Let's keep nouns & adjectives only
#Let's create a function to pull out nouns & adjectives from a string of text

from nltk import word_tokenize, pos_tag

def nouns_adj(text):
    '''Given a string of text, tokenize the text and pull out only the nouns and adjectives.'''
    is_noun_adj = lambda pos: pos[:2] == 'NN' or pos[:2] == 'JJ'
    tokenized = word_tokenize(text)
    nouns_adj = [word for (word, pos) in pos_tag(tokenized) if is_noun_adj(pos)] 
    return ' '.join(nouns_adj)

#We transform our list of content with the function we have just created
sinekCorpusNounsAdj = sinekCorpus.apply(nouns_adj)
sinekCorpusNounsAdj

# Create a new document-term matrix using only nouns
from sklearn.feature_extraction import text 
from sklearn.feature_extraction.text import CountVectorizer

# Re-add the additional stop words since we are recreating the document-term matrix
add_stop_words = ['like', 'im', 'know', 'just', 'dont', 'thats', 'right', 'people',
                  'youre', 'got', 'gonna', 'time', 'think', 'yeah', 'said']
stop_words = text.ENGLISH_STOP_WORDS.union(add_stop_words)

# Recreate a document-term matrix with only nouns
cvna = CountVectorizer(stop_words=stop_words)
data_cvna = cvna.fit_transform(sinekCorpusNounsAdj)
sinekDtmNounsAdj = pd.DataFrame(data_cvna.toarray(), columns=cvna.get_feature_names())
sinekDtmNounsAdj

# Create the gensim corpus
corpusna = matutils.Sparse2Corpus(scipy.sparse.csr_matrix(sinekDtmNounsAdj.transpose()))

# Create the vocabulary dictionary
id2wordna = dict((v, k) for k, v in cvna.vocabulary_.items())

# Let's start with 2 topics
ldana = models.LdaModel(corpus=corpusna, num_topics=2, id2word=id2wordna, passes=10)
ldana.print_topics()

#3 topics
ldana = models.LdaModel(corpus=corpusna, num_topics=3, id2word=id2wordna, passes=10)
ldana.print_topics()

#4 topics
ldana = models.LdaModel(corpus=corpusna, num_topics=4, id2word=id2wordna, passes=10)
ldana.print_topics()
