==> Development/6_TopicModeling.ipynb <==
# Topic modeling

Now we worked on Simon Sinek's posts to understand how to use the gensim library & LDA, let's generalize the method to the entire authors corpus.

## Libraries

import pandas as pd
import pickle
from gensim import matutils, models
import numpy
import scipy.sparse
from sklearn.feature_extraction import text 
from sklearn.feature_extraction.text import CountVectorizer

## Pre-processing

#Import the corpus
data = pd.read_pickle("contentCorpus.pkl")
#Remove useless columns for this analysis. 
data.drop(['#Reactions','#Comments','Location','Followers','Time_spent','Media_type'],axis=1,inplace=True)

#Because we want to define topics across all authors, we don't care about each post individually
#Thus let's group posts content by author
authorCorpus = pd.read_pickle('contentAuthorCorpus.pkl')
authorCorpus.drop(['Followers'],axis=1,inplace=True)

#Transform list of authors into an index
authorCorpus.set_index('Name',inplace=True)

#Let's transform the dataframe into a serie
authorCorpus = authorCorpus.squeeze()

#Let's keep nouns & adjectives only
#Let's create a function to pull out nouns & adjectives from a string of text

from nltk import word_tokenize, pos_tag

def nouns_adj(text):
    '''Given a string of text, tokenize the text and pull out only the nouns and adjectives.'''
    is_noun_adj = lambda pos: pos[:2] == 'NN' or pos[:2] == 'JJ'
    tokenized = word_tokenize(text)
    nouns_adj = [word for (word, pos) in pos_tag(tokenized) if is_noun_adj(pos)] 
    return ' '.join(nouns_adj)

#We transform our list of content with the function we have just created
CorpusNounsAdj = authorCorpus.apply(nouns_adj)
CorpusNounsAdj

# We create a document-term matrix using CountVectorizer, and exclude common English stop words

cv = CountVectorizer(stop_words='english')
dataCv = cv.fit_transform(CorpusNounsAdj)
dataDtm = pd.DataFrame(dataCv.toarray(), columns=cv.get_feature_names_out(),index=CorpusNounsAdj.index)

print(dataDtm.shape)
dataDtm

## 1st attempt

# We're going to put the term-document matrix into a new gensim format, from df --> sparse matrix --> gensim corpus
sparse_counts = scipy.sparse.csr_matrix(dataDtm)
corpus = matutils.Sparse2Corpus(sparse_counts) #It's actually a dtm

# Gensim also requires dictionary of the all terms and their respective location in the term-document matrix
id2word = dict((v, k) for k, v in cv.vocabulary_.items())

# Now that we have the corpus (term-document matrix) and id2word (dictionary of location: term),
# we need to specify two other parameters as well - the number of topics and the number of passes
lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=2, passes=10)
lda.print_topics()

#4topics
lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=4, passes=5)
lda.print_topics()

## 2nd attempt

Let's remove some words from this topics

add_stop_words = ["abadesi","abbi","abdullah","abhishant","aback","aarogyasetu","abeid","aaa","abigail","aams"
                 ,"aberman","abidjan","abhishek","abbass","abhishant","abacusagency","abi","ab","abdi"]
stop_words = text.ENGLISH_STOP_WORDS.union(add_stop_words)

# Recreate a document-term matrix with this correction
cv = CountVectorizer(stop_words=stop_words)
data_cv = cv.fit_transform(CorpusNounsAdj)
dataDtm = pd.DataFrame(data_cv.toarray(), columns=cv.get_feature_names())
dataDtm

# We're going to put the term-document matrix into a new gensim format, from df --> sparse matrix --> gensim corpus
sparse_counts = scipy.sparse.csr_matrix(dataDtm)
corpus = matutils.Sparse2Corpus(sparse_counts) #It's actually a dtm

#4topics
lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=4, passes=10)
lda.print_topics()

Ok.  
So there are a lot of words with no sense.

## 3rd attempt

We will try to remove words with no sense from the corpus.

To do so, we will use the dictionary present in NLTK : nltk.words.

It contains nearly all english words. 


import nltk
nltk.download('words')

englishWordsDic = nltk.corpus.words.words()
print ("There are {} words in this dictionary".format(len(englishWordsDic)))

#Let's create the function to filter only english words
def isEnglish(text):
    '''Given a string of text, tokenize the text and check if each word is an english word.'''
    tokenized = word_tokenize(text)
    englishWord = [word for word in tokenized if (word in englishWordsDic) == True] 
    return ' '.join(englishWord)

Now we have defined the function, we can apply it to the corpus.  
However, because it take a very long time to process, we will focus on two authors content at first and apply LDA to it.

corpusTest = CorpusNounsAdj.iloc[[6,67]]
corpusTest

corpusTestCleaned = corpusTest.apply(isEnglish)
corpusTestCleaned

#We define stop words
stop_words = text.ENGLISH_STOP_WORDS

#Document term matrix
cv = CountVectorizer(stop_words=stop_words)
data_cv = cv.fit_transform(corpusTest)
dataDtm = pd.DataFrame(data_cv.toarray(), columns=cv.get_feature_names())
dataDtm

# Gensim also requires dictionary of the all terms and their respective location in the term-document matrix
id2word = dict((v, k) for k, v in cv.vocabulary_.items())

# We're going to put the term-document matrix into a new gensim format, from df --> sparse matrix --> gensim corpus
sparse_counts = scipy.sparse.csr_matrix(dataDtm)
corpus = matutils.Sparse2Corpus(sparse_counts) #It's actually a dtm

#4topics
lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=4, passes=20)
lda.print_topics()

"abitofoptimism" in englishWordsDic

### Partial conclusions

The LDA algorithm doesn't work.  
It's probably due to the fact that we have only few document (we gather posts by author).  
So let's use each post individually and do LDA on that.

## 4th attempt

#Import the corpus
contentCorpus = pd.read_pickle("contentCorpus.pkl")
#Remove useless columns for this analysis. 
contentCorpus.drop(['Name','#Reactions','#Comments','Location','Followers','Time_spent','Media_type'],axis=1,inplace=True)
contentCorpus

import nltk
from nltk import word_tokenize, pos_tag

nltk.download('words')
englishWordsDic = nltk.corpus.words.words()
print ("There are {} words in this dictionary".format(len(englishWordsDic)))

Let's create the function to filter only english nouns & adjectives

#Let's keep nouns & adjectives only
#Let's create a function to pull out nouns & adjectives from a string of text

def corpusCleaning(text):
    '''Given a string of text, tokenize the text and pull out only the nouns and adjectives belonging to english dictionary.'''
    is_noun_adj = lambda pos: pos[:2] == 'NN' or pos[:2] == 'JJ'
    is_english = lambda word : word in englishWordsDic
    tokenized = word_tokenize(text)
    cleanedtext = [word for (word, pos) in pos_tag(tokenized) if is_noun_adj(pos) and is_english(word)] 
    return ' '.join(cleanedtext)

#Transform the corpus into a Serie
contentCorpus = contentCorpus.squeeze()
contentCorpus

testCorpus = contentCorpus[2:20]
testCorpus

cleanedCorpus = testCorpus.apply(corpusCleaning)
cleanedCorpus

#We define stop words
stop_words = text.ENGLISH_STOP_WORDS
#Document term matrix
cv = CountVectorizer(stop_words=stop_words)
data_cv = cv.fit_transform(cleanedCorpus)
dataDtm = pd.DataFrame(data_cv.toarray(), columns=cv.get_feature_names())
# Gensim also requires dictionary of the all terms and their respective location in the term-document matrix
id2word = dict((v, k) for k, v in cv.vocabulary_.items())
# We're going to put the term-document matrix into a new gensim format, from df --> sparse matrix --> gensim corpus
sparse_counts = scipy.sparse.csr_matrix(dataDtm)
corpus = matutils.Sparse2Corpus(sparse_counts) #It's actually a dtm
#4topics
lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=2, passes=80)
lda.print_topics()

And now for the entire dataset : contentCorpus

cleanedCorpus = contentCorpus.apply(corpusCleaning)
cleanedCorpus

#32 minutes to process (due to "in English Dic")
cleanedCorpus.to_pickle('cleanedPostsCorpus.pkl')

reducedCorpus = cleanedCorpus.iloc[:10000]
reducedCorpus

#We define stop words
stop_words = text.ENGLISH_STOP_WORDS
#Document term matrix
cv = CountVectorizer(stop_words=stop_words)
data_cv = cv.fit_transform(reducedCorpus)
dataDtm = pd.DataFrame(data_cv.toarray(), columns=cv.get_feature_names())
# Gensim also requires dictionary of the all terms and their respective location in the term-document matrix
id2word = dict((v, k) for k, v in cv.vocabulary_.items())
# We're going to put the term-document matrix into a new gensim format, from df --> sparse matrix --> gensim corpus
sparse_counts = scipy.sparse.csr_matrix(dataDtm)
corpus = matutils.Sparse2Corpus(sparse_counts) #It's actually a dtm
# #4topics
lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=1, passes=10)
lda.print_topics()

There is a bug in the LDA algorithm.  
Let's try to figure out what makes this bug occurs !

## 5th attempt

Let's take all authors posts dataset, and apply only Nouns & Adj function on it.  
In other words, we don't use check if a word is in the english dictionary.  
In this way, we will see if our algorithm has a problem.

#Import the corpus
contentCorpus = pd.read_pickle("contentCorpus.pkl")
#Remove useless columns for this analysis. 
contentCorpus.drop(['Name','#Reactions','#Comments','Location','Followers','Time_spent','Media_type'],axis=1,inplace=True)
contentCorpus

import nltk
from nltk import word_tokenize, pos_tag
nltk.download('words')
englishWordsDic = nltk.corpus.words.words()
print ("There are {} words in this dictionary".format(len(englishWordsDic)))

#Let's keep nouns & adjectives only
#Let's create a function to pull out nouns & adjectives from a string of text

def corpusCleaning(text):
    '''Given a string of text, tokenize the text and pull out only the nouns and adjectives belonging to english dictionary.'''
    is_noun_adj = lambda pos: pos[:2] == 'NN' or pos[:2] == 'JJ'
    tokenized = word_tokenize(text)
    cleanedtext = [word for (word, pos) in pos_tag(tokenized) if is_noun_adj(pos)] 
    return ' '.join(cleanedtext)

#Transform the corpus into a Serie
contentCorpus = contentCorpus.squeeze()
contentCorpus

cleanedCorpus = contentCorpus.apply(corpusCleaning)
cleanedCorpus

#Transform the corpus into a Serie
contentCorpus = contentCorpus.squeeze()
contentCorpus
#We define stop words
stop_words = text.ENGLISH_STOP_WORDS
#Document term matrix
cv = CountVectorizer(stop_words=stop_words)
data_cv = cv.fit_transform(cleanedCorpus)
dataDtm = pd.DataFrame(data_cv.toarray(), columns=cv.get_feature_names())
# Gensim also requires dictionary of the all terms and their respective location in the term-document matrix
id2word = dict((v, k) for k, v in cv.vocabulary_.items())
# We're going to put the term-document matrix into a new gensim format, from df --> sparse matrix --> gensim corpus
sparse_counts = scipy.sparse.csr_matrix(dataDtm)
corpus = matutils.Sparse2Corpus(sparse_counts) #It's actually a dtm
#4topics
lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=6, passes=20)
lda.print_topics()

OK the LDA algorithm works correctly without our english words only filter.  
That means we made a mistake in our filter function.  
Let's correct it !

## 6th attempt

#Import the corpus
contentCorpus = pd.read_pickle("contentCorpus.pkl")
#Remove useless columns for this analysis. 
contentCorpus.drop(['Name','#Reactions','#Comments','Location','Followers','Time_spent','Media_type'],axis=1,inplace=True)
contentCorpus

import nltk
from nltk import word_tokenize, pos_tag
nltk.download('words')
englishWordsDic = nltk.corpus.words.words()
print ("There are {} words in this dictionary".format(len(englishWordsDic)))

#Let's keep nouns & adjectives only
#Let's create a function to pull out nouns & adjectives from a string of text

def corpusCleaning(text):
    '''Given a string of text, tokenize the text and pull out only the nouns and adjectives belonging to english dictionary.'''
    is_noun_adj = lambda pos: pos[:2] == 'NN' or pos[:2] == 'JJ'
    is_english = lambda word : word in englishWordsDic
    tokenized = word_tokenize(text)
    NounsAdjList = [word for (word, pos) in pos_tag(tokenized) if is_noun_adj(pos)] 
    #Then, we check if words are english
    englishList = [word for word in NounsAdjList if is_english(word)]
    return ' '.join(englishList)

#Transform the corpus into a Serie
contentCorpus = contentCorpus.squeeze()
contentCorpus

cleanedCorpus = contentCorpus.apply(corpusCleaning)
cleanedCorpus

#It appears that after this cleaning, some strings are empty ""
cleanedCorpus.loc[cleanedCorpus == ''].size

#Therefore we remove these posts
cleaned2Corpus = cleanedCorpus.loc[cleanedCorpus != ""]
cleaned2Corpus.size

#Let's save this df
cleaned2Corpus.to_pickle("cleanedPostsCorpus.pkl")

#We define stop words
stop_words = text.ENGLISH_STOP_WORDS
#Document term matrix
cv = CountVectorizer(stop_words=stop_words)
data_cv = cv.fit_transform(cleaned2Corpus)
dataDtm = pd.DataFrame(data_cv.toarray(), columns=cv.get_feature_names())
# Gensim also requires dictionary of the all terms and their respective location in the term-document matrix
id2word = dict((v, k) for k, v in cv.vocabulary_.items())
# We're going to put the term-document matrix into a new gensim format, from df --> sparse matrix --> gensim corpus
sparse_counts = scipy.sparse.csr_matrix(dataDtm)
corpus = matutils.Sparse2Corpus(sparse_counts) #It's actually a dtm
#4topics
lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=6, passes=20)
lda.print_topics()

print(sparse_counts)
